requirements.txt
selenium
python-dotenv
sqlalchemy
sympy
jsonlines
Werkzeug
packaging
-e git+https://github.com/LCBRU/lbrc_selenium.git@main#egg=lbrc_selenium
